Renamed terrors packages to errosx. Only because it's less weird.
